feat: New Battle update logic

+++ IT/codekatabattle/src/main/java/it/polimi/codekatabattle/entities/TournamentParticipant.java
@@ -20,6 +20,6 @@ public class TournamentParticipant extends BaseEntity {
     private String username;
 
     @Column
-    private float score;
+    private Integer score;
 
 }
